requirements.txt
flask
pypdf2
# langchain
# langchain-community
# langchain-openai==0.0.6
# python-dotenv
# huggingface_hub
langchain==0.1.6
langchain-cli==0.0.21
huggingface_hub==0.21.4
python-dotenv==1.0.0
pydantic==1.10.13
# Flask-Session
InstructorEmbedding
faiss-cpu
sentence-transformers===2.2.2
torch===2.2.0
# langchain-huggingface #for llm 